requirements.txt
torch==1.13.1+cu116
torchvision==0.14.1+cu116
Pillow==9.4.0